# openapi.yaml
openapi: 3.0.3
info:
  title: LangGraph Multi-Agent System API
  description: |
    A comprehensive multi-agent workflow system built with FastAPI and LangGraph.

    This API provides endpoints for managing tasks, agents, workflows, and system operations.
    The system supports multiple execution modes, real-time communication via WebSockets,
    and comprehensive monitoring and health checking capabilities.

    ## Features
    - Task management with priority and type classification
    - Multi-agent coordination and execution
    - Real-time status updates via WebSocket
    - Comprehensive system monitoring and health checks
    - LLM configuration management
    - Batch operations and statistics
    - Meta-agent conversation management

    ## Authentication
    The API uses JWT-based authentication for securing endpoints.

    ## Rate Limiting
    Rate limiting is applied to prevent abuse and ensure system stability.

    ## WebSocket Support
    Real-time communication is supported through WebSocket endpoints for live updates.
  version: 1.0.0
  contact:
    name: LangGraph Multi-Agent Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/meta-agent/conversations:
    post:
      tags:
        - Meta Agent
      summary: Start new conversation
      description: Start a new conversation with the meta-agent
      operationId: startConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - initial_message
              properties:
                initial_message:
                  type: string
                  description: Initial message to start the conversation
                context:
                  type: object
                  description: Additional context for the conversation
                user_id:
                  type: string
                  description: User ID
                priority:
                  type: string
                  enum: [low, medium, high]
                  default: medium
      responses:
        '201':
          description: Conversation started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                  message:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/meta-agent/conversations/{conversation_id}:
    get:
      tags:
        - Meta Agent
      summary: Get conversation
      description: Retrieve details of a specific conversation
      operationId: getConversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      responses:
        '200':
          description: Conversation retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        message_id:
                          type: string
                        role:
                          type: string
                          enum: [user, assistant, system]
                        content:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        metadata:
                          type: object
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
                  status:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Meta Agent
      summary: Delete conversation
      description: Delete a conversation and all associated messages
      operationId: deleteConversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      responses:
        '200':
          description: Conversation deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/meta-agent/conversations/{conversation_id}/messages:
    post:
      tags:
        - Meta Agent
      summary: Add message to conversation
      description: Add a new message to an existing conversation
      operationId: addMessage
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Message content
                role:
                  type: string
                  enum: [user, assistant, system]
                  default: user
                metadata:
                  type: object
                  description: Additional metadata
      responses:
        '201':
          description: Message added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: string
                  conversation_id:
                    type: string
                  role:
                    type: string
                  content:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  response:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/meta-agent/conversations/{conversation_id}/analyze-complexity:
    post:
      tags:
        - Meta Agent
      summary: Analyze conversation complexity
      description: Analyze the complexity of a conversation and provide insights
      operationId: analyzeComplexity
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                analysis_depth:
                  type: string
                  enum: [basic, detailed, comprehensive]
                  default: detailed
                focus_areas:
                  type: array
                  items:
                    type: string
                  description: Specific areas to focus analysis on
      responses:
        '200':
          description: Complexity analysis completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                  complexity_score:
                    type: number
                    minimum: 0
                    maximum: 10
                  complexity_level:
                    type: string
                    enum: [simple, moderate, complex, very_complex]
                  insights:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          type: string
                        description:
                          type: string
                        confidence:
                          type: number
                          minimum: 0
                          maximum: 1
                  recommendations:
                    type: array
                    items:
                      type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/meta-agent/conversations/{conversation_id}/decompose:
    post:
      tags:
        - Meta Agent
      summary: Decompose tasks from conversation
      description: Decompose complex tasks identified in the conversation into smaller subtasks
      operationId: decomposeTasks
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_tasks:
                  type: array
                  items:
                    type: string
                  description: Specific tasks to decompose
                decomposition_strategy:
                  type: string
                  enum: [hierarchical, sequential, parallel, priority_based]
                  default: hierarchical
                max_depth:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 3
      responses:
        '200':
          description: Task decomposition completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                  decomposition:
                    type: array
                    items:
                      type: object
                      properties:
                        task_id:
                          type: string
                        parent_task_id:
                          type: string
                          nullable: true
                        title:
                          type: string
                        description:
                          type: string
                        priority:
                          type: string
                          enum: [low, medium, high, urgent]
                        estimated_duration:
                          type: string
                        dependencies:
                          type: array
                          items:
                            type: string
                        required_capabilities:
                          type: array
                          items:
                            type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Operation success status
        message:
          type: string
          description: Response message
        data:
          type: object
          additionalProperties: true
          nullable: true
          description: Response data
        error_code:
          type: string
          nullable: true
          description: Error code if applicable
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - error_code
        - error_message
      properties:
        success:
          type: boolean
          default: false
          description: Always false for error responses
        error_code:
          type: string
          description: Machine-readable error code
        error_message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: BAD_REQUEST
            error_message: Invalid request parameters
            timestamp: "2024-01-01T00:00:00Z"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: UNAUTHORIZED
            error_message: Authentication required
            timestamp: "2024-01-01T00:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: NOT_FOUND
            error_message: Resource not found
            timestamp: "2024-01-01T00:00:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error_code: INTERNAL_ERROR
            error_message: Internal server error
            timestamp: "2024-01-01T00:00:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
security:
  - BearerAuth: []
...